third_party/WebKit/LayoutTests/fast/notifications/notifications-constructor-request-permission.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test.js"></script>
    </head>
    <body>
        <p>
            Tests that it is possible to request permission to display Web
            Notifications from a document context, for which the passed callback
            will be invoked with the requests' result. It also makes sure that
            not passing a callback won't result in a crash.
        </p>
        <div id="console"></div>
        <script>
            if (window.testRunner) {
                testRunner.dumpAsText();
                testRunner.waitUntilDone();
                testRunner.clearWebNotificationPermissions();
            }

            var result;

            // This should not crash given that the callback is optional.
            Notification.requestPermission();

            // This should log the default permission, since no other permission
            // has been set yet.
            Notification.requestPermission(function (localResult) {
                result = localResult;
                shouldBeEqualToString("result", "default");

                if (!window.testRunner) {
                    testPassed("The remainder of this test requires the TestRunner.");
                    return;
                }

                testRunner.grantWebNotificationPermission("file://", true);
                Notification.requestPermission(function (localResult) {
                    result = localResult;
                    shouldBeEqualToString("result", "granted");

                    testRunner.notifyDone();
                });
            });
        </script>
    </body>
</html>
